requirements.txt
pandas
python-dotenv
roboflow
jupyter
opencv-python
numpy
matplotlib
supervision
